Rename ArborX_DetailsTraits.hpp -> ArborX_Traits.hpp
