+++ src/details/ArborX_DetailsTreeConstruction.hpp
@@ -18,8 +18,8 @@
 #include <ArborX_DetailsMortonCode.hpp> // morton3D
 #include <ArborX_DetailsNode.hpp>
 #include <ArborX_DetailsTags.hpp>
-#include <ArborX_DetailsTraits.hpp>
 #include <ArborX_Macros.hpp>
+#include <ArborX_Traits.hpp>
 
 // FIXME provides definition of Kokkos::Iterate for Kokkos_CopyViews.hpp
 #include <KokkosExp_MDRangePolicy.hpp>
